Replaces Rule Quarm::EnablePetExperienceSplit with EQA::EnableClassicPetExperienceSplit
This update removes experience penalties to groups when a single pet
does > 50% experience. There is no available evidence showing this is
a classic mechanic.  However, it leaves the 50% experience penalty in
place for solo'ing pet classes that allow their pet to do > 50% damage
to a target.

I removed this rule from the Quarm namespace and created an EQA
namespace in order to avoid confusion about the effect of this rule.